Change the functionality of iterating over the myLibrary array
We create the displayBook() func to create the book card and display it on the
screen.

We create a custom event that gets triggered at the end of the file to display
all the books inside the myLibrary array. This event will later be used to
display all the books again once a new book has been added.