Add remove button to the book cards
When the remove button is clicked the book card gets removed from the
display and from the myLibrary array. The displayBooksEvent gets
dispatched when a book gets removed, so the books that are still
left on the myLibrary array, get new indexes and can get remove
effectively in the future.